+++ robotframework-ls/src/robotframework_ls/impl/filesystem_section_completions.py
@@ -232,21 +232,21 @@ def get_requisites(completion_context: ICompletionContext) -> Optional[_Requisit
     if token_info is not None:
         # Library
         token = ast_utils.get_library_import_name_token(
-            token_info.node, token_info.token
+            token_info.node, token_info.token, generate_empty_on_eol=True
         )
         if token is not None:
             return _Requisites(token, "library")
 
         # Resource
         token = ast_utils.get_resource_import_name_token(
-            token_info.node, token_info.token
+            token_info.node, token_info.token, generate_empty_on_eol=True
         )
         if token is not None:
             return _Requisites(token, "resource")
 
         # Variable
         token = ast_utils.get_variables_import_name_token(
-            token_info.node, token_info.token
+            token_info.node, token_info.token, generate_empty_on_eol=True
         )
         if token is not None:
             return _Requisites(token, "variables")
@@ -262,6 +262,19 @@ def complete(completion_context: ICompletionContext) -> List[CompletionItemTyped
         requisites = get_requisites(completion_context)
         if requisites is None:
             return []
+
+        return complete_with_requisites(completion_context, requisites)
+
+    except:
+        log.exception()
+
+    return []
+
+
+def complete_with_requisites(
+    completion_context: ICompletionContext, requisites: _Requisites
+) -> List[CompletionItemTypedDict]:
+    try:
 
         if requisites.is_library:
             return _get_library_completions(completion_context, requisites.token)