+++ robotframework-ls/src/robotframework_ls/impl/robot_specbuilder.py
@@ -192,6 +192,8 @@ class LibraryDoc(object):
         self.scope = scope
         self.named_args = named_args
         self.doc_format = doc_format or "ROBOT"
+        if source and source.startswith("<"):  # Deal with <string>
+            source = None
         self._source = source
         self.lineno = lineno
         self.inits = []
@@ -483,6 +485,8 @@ class KeywordDoc(object):
         self.doc = doc
         self.tags = tags
         self._shortdoc = ""
+        if source and source.startswith("<"):
+            source = None
         self._source = source
         self.lineno = lineno
 
@@ -508,7 +512,9 @@ class KeywordDoc(object):
 
     @property
     def deprecated(self) -> bool:
-        return self.doc.startswith("*DEPRECATED") and "*" in self.doc[1:]
+        from robotframework_ls.impl.text_utilities import has_deprecated_text
+
+        return has_deprecated_text(self.doc)
 
     @property  # type: ignore
     @instance_cache