+++ robotframework-ls/docs/changelog.md
@@ -5,13 +5,25 @@ NEXT
 
 - Code action: Surround with try..except.
 - Code action: Surround with try..except..finally.
-- `robot.libraries.libdoc.blacklist` can be used to blacklist libraries which should not appear in code-completion.
+- `robot.libraries.blacklist` can be used to blacklist libraries which should not appear in code-completion.
 - Interactive Console: (Robot Framework 4 onwards): When a single keyword call is executed and it has a non-None return, its return value is printed.
-- Interactive Console: Typing a variable in the `Interactive Console` shows that variable. [#871](https://github.com/robocorp/robotframework-lsp/issues/871) 
+- Interactive Console: Typing a variable in the `Interactive Console` shows that variable. [#871](https://github.com/robocorp/robotframework-lsp/issues/871)
+- Improve `IF`/`WHILE`/`Run Keyword If` snippet completions to make it a bit clearer that python expressions are expected.
+- Building the model of the flow explorer no longer times out and a progress dialog is shown while building it.
+- Deprecated libraries:
+    - `robot.libraries.deprecated` can be used to mark libraries as deprecated.
+    - Keywords from deprecated libraries will not appear in the auto-import code-completion (so, they'll only be available if the `Library` is added to the `Settings`).
+    - Libraries which start with `*DEPRECATED*` in its doc are also considered deprecated.
+    - Keywords from libraries marked as deprecated will be shown as deprecated.
+- Requesting completions right after `Libraries   ` without any additional name will show completions for all known (pre-loaded) libraries.
+- Completions for variables are shown without having to enter `$` nor `${`.
+
 
 ### Bugfixes
 
-- Properly report about undefined variable in RETURN. [#865](https://github.com/robocorp/robotframework-lsp/issues/865)
+- Properly report about undefined variable in `RETURN`. [#865](https://github.com/robocorp/robotframework-lsp/issues/865)
+- References to variables used in `Evaluate` arguments are now properly collected.
+- Don't show completion for the variable being currently defined in variable assign. 
 
 
 New in 1.8.0 (2023-01-23)